# Conflicts:
#	src/main/java/com/expedienteclinico/expedienteclinico/exceptions/DineroInsuficienteException.java
#	src/main/java/com/expedienteclinico/expedienteclinico/models/Banco.java

+++ src/main/java/com/expedienteclinico/expedienteclinico/repositories/rrhh/ISchedulesRepository.java
@@ -1,6 +1,12 @@
 package com.expedienteclinico.expedienteclinico.repositories.rrhh;
+import com.expedienteclinico.expedienteclinico.models.rrhh.DepartmentsModel;
 import com.expedienteclinico.expedienteclinico.models.rrhh.SchedulesModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISchedulesRepository extends JpaRepository<SchedulesModel, Integer> {
+import java.util.Optional;
+
+public interface ISchedulesRepository extends JpaRepository<SchedulesModel, Long> {
+//    boolean existsOverlap(String name);
+//
+//    Optional<SchedulesModel> findByNameIgnoreCase(String name);
 }
